pyproject.toml
[tool.poetry]
name = "google-keep"
version = "0.1.0"
description = ""
authors = ["Phan Văn Thành <thanhphanvan1610@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.142.0"
gkeepapi = "^0.16.0"
python-dotenv = "^1.0.1"
schedule = "^1.2.2"
groq = "^0.9.0"
python-telegram-bot = "^21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
